src/requirements.txt
langchain==0.3.4
langchain-core==0.3.21
langchain-google-genai==2.0.5
python-dotenv==1.0.1
